requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Data Validation
pydantic==2.9.2

# Database
pyodbc==5.2.0

# CORS & Middleware
python-multipart==0.0.12

pyodbc
passlib[bcrypt]
python-jose[cryptography]
python-dotenv